API Reference

Data configuration

Missing docstring.

Missing docstring for dataset. Check Documenter's build log for details.

DataSets.DataSet — Type

A DataSet is a metadata overlay for data held locally or remotely which is unopinionated about the underlying storage mechanism.

The data in a DataSet has a type which implies an index; the index can be used to partition the data for processing.

DataSets.DataProject — Type

DataProject

A data project is a collection of DataSets with associated names. Names are unique within the project.

Missing docstring.

Missing docstring for DataSets.load_project. Check Documenter's build log for details.

Data function entry points

DataSets.@datafunc — Macro

@datafunc function f(x::DT=>T, y::DS=>S...)
    ...
end

Define the function f(x::T, y::S, ...) and add data dispatch rules so that f(x::DataSet, y::DataSet) will open datasets matching dataset types DT,DS as Julia types T,S.

DataSets.@datarun — Macro

@datarun [proj] func(args...)

Run func with the named DataSets from the list args.

Example

Load DataSets named a,b as defined in Data.toml, and pass them to f().

proj = DataSets.load_project("Data.toml")
@datarun proj f("a", "b")

Data Models

Missing docstring.

Missing docstring for Blob. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BlobTree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for newfile. Check Documenter's build log for details.

Missing docstring.

Missing docstring for newdir. Check Documenter's build log for details.

Storage Drivers

DataSets.add_storage_driver — Function

add_storage_driver(driver_name=>storage_opener)

Associate DataSet storage driver named driver_name with storage_opener. When a dataset with storage.driver == driver_name is opened, storage_opener(user_func, storage_config, dataset) will be called. Any existing storage driver registered to driver_name will be overwritten.

As a matter of convention, storage_opener should generally take configuration from storage_config which is just dataset.storage. But to avoid config duplication it may also use the content of dataset, (for example, dataset.uuid).

Packages which define new storage drivers should generally call add_storage_driver() within their __init__() functions.